requirements.txt
django==4.2.11 
djangorestframework==3.15.1 
psycopg2-binary==2.9.9
gunicorn==22.0.0